Moved application from 'api' package to the main one to simplify the startup and beans registration
