+++ samples/event-sourcing-esdb-simple/src/main/java/io/eventdriven/ecommerce/ECommerceApplication.java
@@ -1,0 +1,11 @@
+package io.eventdriven.ecommerce;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ECommerceApplication {
+  public static void main(String[] args) {
+    SpringApplication.run(ECommerceApplication.class, args);
+  }
+}
